refactor : LineCreator 인터페이스 및 구현체 추가로 인한 리팩터링
