+++ src/main/java/ladder/domain/JerryLine.java
@@ -6,18 +6,20 @@ import ladder.engine.Line;
 import java.util.ArrayList;
 import java.util.List;
 
+import static ladder.domain.JerryLineCreator.MIN_WIDTH;
+
 public class JerryLine implements Line {
     private List<Direction> points = new ArrayList<>();
-    public static final int MIN_WIDTH = 2;
+//    public static final int MIN_WIDTH = 2;
 
     public JerryLine(List<Direction> points) {
-        validateWidth(points.size());
+//        validateWidth(points.size());
 
         this.points = points;
     }
 
     public static JerryLine createLineWithWidth(int width) {
-        validateWidth(width);
+//        validateWidth(width);
 
         List<Direction> points = new ArrayList<>();
 
@@ -30,11 +32,11 @@ public class JerryLine implements Line {
         return new JerryLine(points);
     }
 
-    private static void validateWidth(int points) {
-        if (points < MIN_WIDTH) {
-            throw new IllegalArgumentException("사다리 폭은 " + MIN_WIDTH + " 이상이어야 합니다.");
-        }
-    }
+//    private static void validateWidth(int points) {
+//        if (points < MIN_WIDTH) {
+//            throw new IllegalArgumentException("사다리 폭은 " + MIN_WIDTH + " 이상이어야 합니다.");
+//        }
+//    }
 
     public List<Direction> getPoints() {
         return points;
