+++ src/main/java/ladder/engine/LineCreator.java
@@ -1,4 +1,7 @@
 package ladder.engine;
 
+import ladder.domain.JerryLine;
+
 public interface LineCreator {
+    Line createLine(int width);
 }
